move all matching and processing to individual files
