+++ src/tokenizer/lexer/token_processors/bang.rb
@@ -1,0 +1,13 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def bang?(chunk)
+        chunk =~ /^[#{BANG}]$/
+      end
+
+      def process_bang(_chunk)
+        (@tokens << Token.new(Token::BANG)).last
+      end
+    end
+  end
+end
