+++ src/tokenizer/lexer/token_processors/break.rb
@@ -1,0 +1,14 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def break?(chunk)
+        chunk =~ /^(終|お)わり$/
+      end
+
+      def process_break(_chunk)
+        validate_scope Scope::TYPE_LOOP, ignore: [Scope::TYPE_IF_BLOCK]
+        (@tokens << Token.new(Token::BREAK)).last
+      end
+    end
+  end
+end
