+++ src/tokenizer/lexer/token_processors/comp_2_lteq.rb
@@ -1,0 +1,14 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def comp_2_lteq?(chunk)
+        chunk =~ /.+以下$/
+      end
+
+      def process_comp_2_lteq(chunk)
+        @stack << comp_token(chunk.chomp('以下'))
+        Token.new Token::COMP_2_LTEQ
+      end
+    end
+  end
+end
