+++ src/tokenizer/lexer/token_processors/comp_2_yori.rb
@@ -1,0 +1,14 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def comp_2_yori?(chunk)
+        chunk =~ /.+より$/
+      end
+
+      def process_comp_2_yori(chunk)
+        @stack << comp_token(chunk.chomp('より'))
+        Token.new Token::COMP_2_YORI
+      end
+    end
+  end
+end
