+++ src/tokenizer/lexer/token_processors/comp_3_neq.rb
@@ -1,0 +1,13 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def comp_3_neq?(chunk)
+        chunk =~ /^(等|ひと)しくなければ$/
+      end
+
+      def process_comp_3_neq(_chunk)
+        close_if_statement [Token.new(Token::COMP_NEQ)]
+      end
+    end
+  end
+end
