+++ src/tokenizer/lexer/token_processors/comp_3_not.rb
@@ -1,0 +1,13 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def comp_3_not?(chunk)
+        chunk == 'でなければ'
+      end
+
+      def process_comp_3_not(chunk)
+        process_comp_3 chunk, reverse?: true
+      end
+    end
+  end
+end
