+++ src/tokenizer/lexer/token_processors/else_if.rb
@@ -1,0 +1,15 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def else_if?(chunk)
+        chunk =~ /^(もしくは|または)$/
+      end
+
+      def process_else_if(_chunk)
+        raise Errors::UnexpectedElseIf unless @context.inside_if_block?
+        @context.inside_if_condition = true
+        (@tokens << Token.new(Token::ELSE_IF)).last
+      end
+    end
+  end
+end
