+++ src/tokenizer/lexer/token_processors/function_call.rb
@@ -1,0 +1,29 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def function_call?(chunk)
+        @current_scope.function?(chunk, signature_from_stack) && (
+          @last_token_type == Token::EOL                               ||
+          (@last_token_type == Token::PARAMETER && !parameter?(chunk)) ||
+          ((@last_token_type == Token::IF || @last_token_type == Token::ELSE_IF) && question?(@reader.peek_next_chunk))
+        )
+      end
+
+      def process_function_call(chunk)
+        destination = @context.inside_if_condition? ? @stack : @tokens
+
+        signature = signature_from_stack
+        function = @current_scope.get_function chunk, signature
+
+        function_call_parameters_from_stack(function).each { |t| destination << t }
+
+        token = Token.new(
+          Token::FUNCTION_CALL,
+          function[:name],
+          sub_type: function[:built_in?] ? Token::FUNC_BUILT_IN : Token::FUNC_USER
+        )
+        (destination << token).last
+      end
+    end
+  end
+end
