+++ src/tokenizer/lexer/token_processors/if.rb
@@ -1,0 +1,14 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      def if?(chunk)
+        chunk == 'もし'
+      end
+
+      def process_if(_chunk)
+        @context.inside_if_condition = true
+        (@tokens << Token.new(Token::IF)).last
+      end
+    end
+  end
+end
