+++ src/tokenizer/lexer/token_processors/rvalue.rb
@@ -1,0 +1,24 @@
+module Tokenizer
+  class Lexer
+    module TokenProcessors
+      # An rvalue is either a primitive, special identifier, or scoped variable.
+      def rvalue?(chunk)
+        value?(chunk) || @current_scope.variable?(chunk)
+      end
+
+      # TODO: (v1.1.0) Cannot assign keys / indices to themselves. (Fix at same time as process_attribute)
+      # No need to validate variable_type because the matcher checks either
+      # primitive or existing variable.
+      def process_rvalue(chunk)
+        chunk = sanitize_variable chunk
+        token = Token.new Token::RVALUE, chunk, sub_type: variable_type(chunk)
+
+        @stack << token
+
+        try_assignment_close
+
+        token
+      end
+    end
+  end
+end
